requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.51
botocore==1.27.51
Django==4.1
django-cors-headers==3.13.0
django-storages==1.13.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
uuid==1.30
